+++ src/devops/files/__init__.py
@@ -2,7 +2,13 @@
 
 from pathlib import Path
 
-from .files import FileType, determine_file_type, get_files_in_dirs, get_staged_files
+from .files import (
+    FileType,
+    determine_file_type,
+    file_exist,
+    get_files_in_dirs,
+    get_staged_files,
+)
 
 __EXECUTION_DIR__ = Path.cwd()
 
@@ -10,6 +16,7 @@ __all__ = [
     "__EXECUTION_DIR__",
     "FileType",
     "determine_file_type",
+    "file_exist",
     "get_files_in_dirs",
     "get_staged_files",
 ]