+++ CHANGELOG.md
@@ -4,12 +4,13 @@ All notable changes to this project will be documented in this file.
 
 ## Next Release
 
-### Features
-
 ### API
 
 - Add cli command `get_latest_tag`
 - Add cli command `increase_latest_tag`
+- Add new license checking rule to `cpp_checks`
+
+### Features
 
 #### Git
 
@@ -24,6 +25,10 @@ All notable changes to this project will be documented in this file.
     global_level = "INFO"
     cpp_level = "DEBUG"
     ```
+
+#### CPP Rules
+
+- Add license check rule for cpp header and source files
 
 ### Deployment